Barnes' double Gamma function

This package provides an implementation of the multiple $\Gamma$ functions $G, \Gamma_2, \Gamma_b$ as defined in this wikipedia article.

BarnesDoubleGamma.logdoublegamma — Function

    logdoublegamma(w, β; tol=10^(-precision(z)))

Compute the logarithm of the double gamma function $Γ_β(w, β)$. with precision tol

BarnesDoubleGamma.doublegamma — Function

    doublegamma(w, β; tol=10^(-precision(z)))

Compute the double gamma function $Γ_β(w)$. with precision tol

BarnesDoubleGamma.loggamma2 — Function

loggamma2(w, β; tol=10^(-precision(z)))

Logarithm of the $Γ_2(w, β)$ function. For efficiency, when using a high precision w it is recommended to set the tolerance to something lower than the precision of w.

BarnesDoubleGamma.gamma2 — Function

gamma2(w, β; tol=10^(-precision(z)))

$Γ_2(w, β)$ function. For efficiency, when using a high precision z it is recommended to set the tolerance to something lower than the precision of z.

BarnesDoubleGamma.log_barnesdoublegamma — Function

log_barnesdoublegamma(z, τ; tol = 10^(-precision(z)))

Logarithm of Barne's G-function $\log(G(z; τ))$, up to a given tolerance. The tolerance is rather pessimistic. For efficiency, when using a high precision z it is recommended to set the tolerance to something lower than the precision of z.

Examples

julia> z = 1; τ = sqrt(3); log_barnesdoublegamma(z, τ)
4.7036054521762405e-12

julia> z = sqrt(big"2"); τ = sqrt(big"3"); log_barnesdoublegamma(z, τ)
0.293394920968626643183216869255154162603276161914057065216580158409738193128475

BarnesDoubleGamma.barnesdoublegamma — Function

barnesdoublegamma(z, τ; tol=10^(-precision(z)))

Barne's G-function $G(z, τ)$, up to a given tolerance. For efficiency, when using a high precision z it is recommended to set the tolerance to something lower than the precision of z.

Examples

julia> z = big"1"; τ = sqrt(big"3"); barnesdoublegamma(z, τ)
0.9999999999999999999999999999999999999999999345543169494792740822465202042518853

julia> z = sqrt(big"2"); τ = sqrt(big"3"); barnesdoublegamma(z, τ)
1.340972263940081256497568500074283394055091669857109294097794355305731639439863

julia> z = big(sqrt(3)); τ = sqrt(big"3"); barnesdoublegamma(z, τ)
1.488928335365086422942328604671778776079655470648475974888882080914824321973775