Barnes' double Gamma function

This package provides an implementation of the multiple $\Gamma$ functions $G, \Gamma_2, \Gamma_b$ as defined in this wikipedia article.

Algorithm

Internally, the double gamma functions are computed using a product formula from this paper:

\[G(z ; \tau)=\frac{1}{\tau \Gamma(z)} e^{\tilde{a}(\tau) \frac{z}{\tau}+\tilde{b}(\tau) \frac{z^2}{2 \tau^2}} \prod_{m \geq 1} \frac{\Gamma(m \tau)}{\Gamma(z+m \tau)} e^{z \psi(m \tau)+\frac{z^2}{2} \psi^{\prime}(m \tau)},\]

where

\[\begin{aligned} C(\tau) & :=\sum_{k=1}^{m-1} \psi(k \tau)+\frac{1}{2} \psi(m \tau)-\frac{1}{\tau} \ln \left(\frac{\Gamma(m \tau)}{\sqrt{2 \pi}}\right)-\frac{\tau}{12} \psi^{\prime}(m \tau) \\ & +\frac{\tau^3}{720} \psi^{(3)}(m \tau)-\frac{\tau^5}{30240} \psi^{(5)}(m \tau)+\frac{\tau^7}{1209600} \psi^{(7)}(m \tau)+O\left(m^{-9}\right) \\ D(\tau) & :=\sum_{k=1}^{m-1} \psi^{\prime}(k \tau)+\frac{1}{2} \psi^{\prime}(m \tau)-\frac{1}{\tau} \psi(m \tau)-\frac{\tau}{12} \psi^{\prime \prime}(m \tau) \\ & +\frac{\tau^3}{720} \psi^{(4)}(m \tau)-\frac{\tau^5}{30240} \psi^{(6)}(m \tau)+\frac{\tau^7}{1209600} \psi^{(8)}(m \tau)+O\left(m^{-10}\right) . \end{aligned}\]

and

\[\tilde{a}(\tau)=a(\tau)-\gamma \tau=\frac{\tau}{2} \ln (2 \pi \tau)+\frac{1}{2} \ln (\tau)-\tau C(\tau), \quad \tilde{b}(\tau)=b(\tau)+\frac{\pi^2 \tau^2}{6}=-\tau \ln (\tau)-\tau^2 D(\tau),\]

the functions $\Gamma$ and $\psi$ are the Euler gamma and digamma function:

\[\psi(z) = (\log \Gamma)'(z).\]

Functions exported by the module

BarnesDoubleGamma.logdoublegamma — Function

    logdoublegamma(w, β)

Compute the logarithm of the double gamma function $Γ_β(w, β)$.

BarnesDoubleGamma.doublegamma — Function

    doublegamma(w, β)

Compute the double gamma function $Γ_β(w)$.

BarnesDoubleGamma.log_barnesdoublegamma — Function

log_barnesdoublegamma(z, τ)

Logarithm of Barne's G-function $\log(G(z; τ))$. Can get very expensive for high precision.

Examples

julia> z = 1; τ = sqrt(3); log_barnesdoublegamma(z, τ) ≈ -3.5564013784958066e-9
true

julia> z = sqrt(big"2"); τ = sqrt(big"3"); log_barnesdoublegamma(z, τ) ≈ 0.293394920968626643183216869255154162603276275448888004730390602371621786480874
true

BarnesDoubleGamma.barnesdoublegamma — Function

barnesdoublegamma(z, τ)

Barne's G-function $G(z, τ)$. Can get very expensive for high precision.

Examples

julia> z = big"1"; τ = sqrt(big"3"); barnesdoublegamma(z, τ) ≈ 1
true

julia> z = sqrt(big"2"); τ = sqrt(big"3"); barnesdoublegamma(z, τ) ≈ 1.340972263940081256497568500074283394055091822104168575112011391955491855627026
true

julia> s3 = sqrt(big"3"); z = s3; τ = s3; barnesdoublegamma(z, τ) ≈ (2big(π))^((τ-1)/2)/sqrt(τ)
true

BarnesDoubleGamma.loggamma2 — Function

loggamma2(w, β)

Logarithm of the $Γ_2(w, β)$ function.

BarnesDoubleGamma.gamma2 — Function

gamma2(w, β)

$Γ_2(w, β)$ function.